sync backend/ do monorepo (94ad96b)
  73030d3 fix(competencias): unidade com matriz aparece na avaliacao do gestor

+++ src/main/java/br/jus/tjgo/kaizen/controller/CompetenciasGestorController.java
@@ -211,7 +211,7 @@ public class CompetenciasGestorController {
     // Ids das unidades que já têm a Matriz de Competências (do tipo) validada — usado para
     // filtrar o seletor de unidade da avaliação (só faz sentido avaliar onde há referencial).
     @GetMapping("/unidades-com-matriz")
-    public List<Long> unidadesComMatriz(
+    public List<Integer> unidadesComMatriz(
             @RequestParam(value = "tipo", required = false, defaultValue = "gestor") String tipo) {
         return service.unidadesComMatrizValidada(tipo);
     }
